http://wordpress.org/support/topic/warning-invalid-argument-supplied-for-foreach-14?replies=8
Hey aehageman

I've found out some data on a couple of issues and corrected them, but am still moving through this thread. 

What caused the errors from the first two posts was the creation of a Split Testing Group without adding landing pages to it. 

I added a restriction to not attempt to create a group if no landers are included. 

git-svn-id: https://plugins.svn.wordpress.org/landing-pages/trunk@657769 b8457f37-d9ea-0310-8a92-e5e31aec5664